chore: move dts plugins outside of builder
