+++ package.json
@@ -7,11 +7,12 @@
     "build": "webpack --node-env production",
     "components": "webpack -c webpack.components.config.js",
     "test": "jest -c tests/spec/config.js",
-    "lint": "eslint -c .eslintrc.js packages --ext ts",
+    "lint": "eslint packages --ext ts",
     "resolve": "yarn workspace @abux/resolve",
     "builder": "yarn workspace @abux/builder"
   },
   "dependencies": {
+    "@abux/webpack-dts": "workspace:^",
     "@changesets/cli": "^2.26.1",
     "typescript": "^5.0.4",
     "webpack": "^5.80.0"
@@ -19,6 +20,7 @@
   "devDependencies": {
     "@types/fs-extra": "^11.0.1",
     "@types/jest": "^29.5.1",
+    "@types/node": "^20.3.3",
     "@typescript-eslint/eslint-plugin": "^5.59.1",
     "@typescript-eslint/parser": "^5.59.1",
     "eslint": "^8.39.0",
