+++ packages/builder/src/filters/initConfigs.spec.ts
@@ -1,7 +1,6 @@
 import { type IWebpackConfigs } from '../interfaces'
 import { extractPattern } from '../lib/data'
 import { getMockConfig } from '../../tests/mocks/mockConfigs'
-import { resolveOptions } from '../lib/entries'
 
 jest.mock('../lib/entries', () => {
   return {
@@ -9,9 +8,6 @@ jest.mock('../lib/entries', () => {
     resolveOptions: jest.fn().mockImplementation((p, o) => o(p))
   }
 })
-const mockLibEntries = {
-  resolveOptions
-}
 
 const webEntries = {
   'web/index.ts': {
