--- packages/builder/src/lib/dts/index.js
@@ -1,1 +1,0 @@
-/*! Copyright (c) 2023 ABux. Under MIT license found in the LICENSE file */(()=>{"use strict";var e={574:function(e,t,r){var i=this&&this.__awaiter||function(e,t,r,i){return new(r||(r=Promise))((function(o,s){function n(e){try{a(i.next(e))}catch(e){s(e)}}function l(e){try{a(i.throw(e))}catch(e){s(e)}}function a(e){var t;e.done?o(e.value):(t=e.value,t instanceof r?t:new r((function(e){e(t)}))).then(n,l)}a((i=i.apply(e,t||[])).next())}))},o=this&&this.__asyncValues||function(e){if(!Symbol.asyncIterator)throw new TypeError("Symbol.asyncIterator is not defined.");var t,r=e[Symbol.asyncIterator];return r?r.call(e):(e="function"==typeof __values?__values(e):e[Symbol.iterator](),t={},i("next"),i("throw"),i("return"),t[Symbol.asyncIterator]=function(){return this},t);function i(r){t[r]=e[r]&&function(t){return new Promise((function(i,o){(function(e,t,r,i){Promise.resolve(i).then((function(t){e({value:t,done:r})}),t)})(i,o,(t=e[r](t)).done,t.value)}))}}},s=this&&this.__importDefault||function(e){return e&&e.__esModule?e:{default:e}};Object.defineProperty(t,"__esModule",{value:!0}),t.DtsFilterWriter=t.DtsWriter=t.Dts=void 0;const n=r(7),l=r(828),a=s(r(361)),d=r(285),u=r(584),c="\n";class h extends a.default{generate({name:e,main:t,inputDir:r,projectPath:o,outputPath:s,files:l=[],references:a=[],filePatterns:c=[]}){var h;return i(this,void 0,void 0,(function*(){let i={};const p=(0,u.resolver)(r),f=s?(0,u.resolver)(s).dir():p;if(s=s||p.resolve("index.d.ts"),this.emit("log",`[dts] generate for ${r} > ${s}`),!l.length||o){const e=yield this.getTsConfig(r,o);l=e.fileNames,i=e.compilerOptions}i.declaration=!0,i.target=i.target||n.ScriptTarget.Latest,i.outDir=i.outDir||f.rootPath;const v=p.rootPath||i.rootDir||o;if(!v)throw Error("[dts] input dir path is required");const g=(0,u.resolver)(v),x=f.rootPath||i.outDir,w=l,M=[`baseDir = "${g.rootPath}"`,`target = ${i.target.toString()}`,`outDir = ${x||""}`,`rootDir = ${i.rootDir||""}`,`moduleResolution = ${(null===(h=i.moduleResolution)||void 0===h?void 0:h.toString())||""}`,"files =",...w.map((e=>`  ${e}`))];this.emit("log:verbose","[dts] params:\n"+M.map((e=>`  ${e}`)).join("\n")),yield(0,d.mkdirp)((0,u.getDir)(s));const $=new m({outputPath:s,name:e,main:t,references:a},{filePatterns:c});$.on("log",(e=>this.emit("log",e))),$.on("log:verbose",(e=>this.emit("log:verbose",e))),yield $.write(g.rootPath,i,w)}))}getTsConfig(e,t){var r,s,n,l;return i(this,void 0,void 0,(function*(){const i=(0,u.resolver)(e),a=[t&&(0,u.resolver)(t).resolve("tsconfig.json"),t,i.resolve("tsconfig.json")].filter(Boolean);try{for(var c,h=!0,p=o(a);c=yield p.next(),!(r=c.done);){l=c.value,h=!1;try{const e=l;if(yield(0,d.pathExists)(e))return this.emit("log",`[dts] tsconfig from ${e}`),yield g(e)}finally{h=!0}}}catch(e){s={error:e}}finally{try{h||r||!(n=p.return)||(yield n.call(p))}finally{if(s)throw s.error}}throw Error(`Can't find tsconfig in ${t||i.rootPath}`)}))}}t.Dts=h;class p extends a.default{constructor(e){super(),this.ident="  ",this.mainId="",this.resolvedMainId="",this.externalModules=[],this.options=Object.assign({references:[],excludedPatterns:["**/node_modules/**/*.d.ts","**/.yarn/**/*.d.ts"]},e),this.options.main||(this.options.main="index"),this.on("done",(()=>{this.dispose()}))}write(e,t,r){return i(this,void 0,void 0,(function*(){this.mainId="",this.resolvedMainId="",this.externalModules=[],this.emit("start"),this.emit("log","[dtsw] start");const i=(0,n.createCompilerHost)(t),o=(0,n.createProgram)(r,t,i),s=o.getSourceFiles(),a=(0,u.resolver)(e);let d=[];this.mainId=this.resolveMainId(),this.inDir=a,this.listExternals(s),this.emit("log","[dtsw] process files"),this.writeReferences(),s.some((e=>{var r;const i=(0,u.normalize)(e.fileName);if(!a.includes(i))return this.emit("log:verbose",`[dtsw] process: ignored library ${i}`),!1;if(null===(r=this.options.excludedPatterns)||void 0===r?void 0:r.some((e=>(0,l.minimatch)(i,e))))return this.emit("log:verbose",`[dtsw] process: excluded ${i}`),!1;if(".d.ts"===i.slice(-5))return this.emit("log:verbose",`[dtsw] process: d.ts ${i}`),this.writeDeclaration(e),!1;const s=this.resolveModule({currentModule:a.relative(w(i))});this.mainId===s&&(this.resolvedMainId=s,this.emit("log:verbose",`[dtsw] main found ${s}`),d=this.getModuleExports(e));const c=o.emit(e,((e,r)=>{".d.ts"===e.slice(-5)?(this.emit("log:verbose",`[dtsw] process: ts ${i}`),this.writeDeclaration((0,n.createSourceFile)(i,r,t.target,!0))):this.emit("log:verbose",`[dtsw] process: ignored d.ts ${i}`)}));if(c.emitSkipped||c.diagnostics.length>0)throw this.emit("log:verbose",`[dtsw] process: ts ${i} error`),x(c.diagnostics.concat(o.getSemanticDiagnostics(e)).concat(o.getSyntacticDiagnostics(e)).concat(o.getDeclarationDiagnostics(e)));return!1})),this.writeMainDeclaration(t.target,d),this.done()}))}listExternals(e){this.emit("log","[dtsw] list externals"),e.forEach((e=>{v(e,(e=>{if(f.isModuleDeclaration(e)){const t=e.name;f.isStringLiteral(t)&&this.externalModules.push(t.text)}}))})),this.externalModules.length?this.emit("log:verbose",["[dtsw] list externals:",...this.externalModules.map((e=>`  - ${e}`))].join("\n")):this.emit("log:verbose","[dtsw] list externals: no externals found")}writeReferences(){var e;const t=/^\./;null===(e=this.options.references)||void 0===e||e.forEach((e=>{t.test(e)?(this.emit("log",`[dtsw] ref.path ${e}`),this.writeOutput(`/// <reference path="${e}" />`)):(this.emit("log",`[dtsw] ref.types ${e}`),this.writeOutput(`/// <reference types="${e}" />`))}))}writeDeclaration(e){var t,r;const i=(0,u.resolve)(e.fileName),o="/"+w((null===(t=this.inDir)||void 0===t?void 0:t.relative(i))||"");if(!o)throw Error(`[dtsw] unable to resolve current module for ${o}`);e.externalModuleIndicator?this.writeExternalDeclaration(e,o):i!==(null===(r=this.output)||void 0===r?void 0:r.path)?(this.emit("log:verbose",`[dtsw] declare ${o} from text`),this.writeOutputModule(o,e.text),this.emit("log:verbose",`[dtsw] declare ${o} done`)):this.emit("log:verbose",`[dtsw] declare ignored ${o}`)}writeMainDeclaration(e,t=[]){const r=this.resolvedMainId,i=[];if(t.length&&r){if(this.emit("log:verbose",`[dtsw] declare:main ${r}`),!e||e<n.ScriptTarget.ES2015)this.emit("log:verbose","[dtsw] declare:main require"),i.push(`import main = require('${r}');`),i.push("export = main;");else{t.includes("default")&&(this.emit("log:verbose","[dtsw] declare:main export default"),i.push(`export { default } from '${r}';`));t.some((e=>"default"!==e))&&(this.emit("log:verbose","[dtsw] declare:main export *"),i.push(`export * from '${r}';`))}i.length||this.emit("log:verbose","[dtsw] declare:main no valid exports"),this.writeOutputModule(this.options.name,i.join(c))}}getModuleExports(e){const t=[];return(0,n.forEachChild)(e,(e=>{var r;f.isExportAssignment(e)?t.push("default"):f.isExportDeclaration(e)?t.push("*"):f.isNamedExports(e)?e.elements.forEach((e=>{var r;t.push((null===(r=e.propertyName)||void 0===r?void 0:r.getText())||e.name.getText())})):f.isVariableStatement(e)&&(null===(r=e.modifiers)||void 0===r?void 0:r.some((e=>e.kind===n.SyntaxKind.ExportKeyword)))&&e.declarationList.declarations.length&&t.push("*")})),t}writeExternalDeclaration(e,t){const r=this.resolveModule({currentModule:t});this.emit("log:verbose",`[dtsw] declare:external ${r} (${e.fileName})`);const i=v(e,(e=>{if(f.isExternalModuleReference(e)){const i=e.expression,o=this.resolveImport({importedModule:i.text,currentModule:t});return this.emit("log:verbose",`[dtsw] declare:external ${r}: require ${o}`),` require('${o}')`}if(f.isDeclareKeyWord(e))return this.emit("log:verbose",`[dtsw] declare:external ${r}: ignored declare keyword`),"";if(f.isStringLiteral(e)&&e.parent&&(f.isExportDeclaration(e.parent)||f.isImportDeclaration(e.parent))){const i=e.text,o=this.resolveImport({importedModule:i,currentModule:t});if(o)return this.emit("log:verbose",`[dtsw] declare:external ${r}: import ${o}`),` '${o}'`}}));this.writeOutputModule(r,i.join("")),this.emit("log:verbose",`[dtsw] declare:external ${r} done`)}writeOutput(e){this.output.write(e+c)}writeOutputModule(e,t){const r=this.filterOutput(t);r.length&&(this.emit("log",`[dtsw] declared module ${e}`),this.writeOutput(`declare module '${e}' {`),this.writeOutput(this.deduplicateLines(r).join(c)),this.writeOutput("}"))}deduplicateLines(e){return e}filterOutput(e){return e.split(/[\r\n]+|; |;$/).filter((e=>e&&"export {};"!==e)).map((e=>e.replace(/\t/g,this.ident))).map((e=>`${this.ident}${e}`.replace(/\s{4}/g,this.ident).replace(/ (\w+)\(/,((e,t)=>"import"!==t?` ${t} (`:e)).replace(/^\s+(private|protected) .+$/,""))).filter(Boolean)}resolveModule(e){let t=e.currentModule;return t=this.options.resolvedModule?this.options.resolvedModule(e)||t:this.resolveModuleDefault(t),t=this.prefixModule(t),this.emit("log:verbose",`[dtsw] resolve ${t} (${e.currentModule})`),t}resolveImport(e){const t=this.externalModules.includes(e.importedModule)||!/^\./.test(e.importedModule),r=t?e.importedModule:(0,u.merge)((0,u.getDir)(e.currentModule),e.importedModule);let i=r;return i=this.options.resolvedImport?this.options.resolvedImport({currentModule:e.currentModule,importedModule:r,isExternal:t})||i:this.resolveModuleDefault(i),i=this.prefixModule(i,t),this.emit("log:verbose",`[dtsw] resolve:import ${i}${t?" (external)":""} (${e.currentModule}, ${e.importedModule})`),i}resolveModuleDefault(e){const t=e.replace(/^.+\/src/,"src"),r=t.replace(/\/index$/,"");return this.mainId!==this.prefixModule(t)&&r!==this.options.name?r:t}resolveMainId(){var e;return`${this.options.name}/${(null===(e=this.options.main)||void 0===e?void 0:e.replace(/^\/+/,""))||"index"}`}prefixModule(e,t=!1){return t?e.replace(/^\//,""):`${this.options.name}/${e.replace(/^\//,"")}`}get output(){return this._output||(this._output=(0,d.createWriteStream)(this.options.outputPath,{mode:parseInt("644",8)})),this._output}dispose(){var e;null===(e=this.output)||void 0===e||e.close()}done(){this.emit("log","[dtsw] done"),this.emit("done")}}t.DtsWriter=p;class m extends p{constructor(e,t){super(e),this.filters=t,this.modulePathMap={},this.moduleDepsMap={},this.cachedOutputs={},this.collectModuleDeps=e=>{const t=this.modulePathMap[e];if(!t)return[];const r=this.moduleDepsMap[t]||[],i=r.map((e=>this.collectModuleDeps(e))).flat();return r.concat(i)}}done(){this.emitOutput(),super.done()}emitOutput(){var e;const t=(null===(e=this.filters.filePatterns)||void 0===e?void 0:e.map((e=>`**/${w(e)}`)))||[],r=Object.keys(this.cachedOutputs),i=r.map((e=>this.modulePathMap[e]));let o=r;t.length&&(o=[],i.forEach(((e,i)=>{e&&t.some((t=>(0,l.minimatch)(e,t)))&&o.push(r[i])})),o=o.map((e=>[e].concat(this.collectModuleDeps(e)))).flat().filter(((e,t,i)=>i.indexOf(e)===t&&r.includes(e))).reverse());this.resolvedMainId&&!o.includes(this.options.name)&&o.push(this.options.name),o.forEach((e=>{this.emit("log",`[dtsw] declared module ${e}`),this.writeOutput(this.cachedOutputs[e])}))}writeOutputModule(e,t){var r;const i=this.filterOutput(t);if(!i.length)return;const o=[`declare module '${e}' {`,this.deduplicateLines(i).join(c),"}"].join(c);(null===(r=this.filters.filePatterns)||void 0===r?void 0:r.length)?this.cachedOutputs[e]=o:this.writeOutput(o)}resolveModule(e){const t=super.resolveModule(e);return this.modulePathMap[t]=e.currentModule,t}resolveImport(e){const t=super.resolveImport(e);return this.moduleDepsMap[e.currentModule]||(this.moduleDepsMap[e.currentModule]=[]),this.moduleDepsMap[e.currentModule].push(t),t}}t.DtsFilterWriter=m;const f={isDeclareKeyWord:e=>e&&e.kind===n.SyntaxKind.DeclareKeyword,isImportDeclaration:e=>e&&e.kind===n.SyntaxKind.ImportDeclaration,isExternalModuleReference:e=>e&&e.kind===n.SyntaxKind.ExternalModuleReference,isStringLiteral:e=>e&&e.kind===n.SyntaxKind.StringLiteral,isExportDeclaration:e=>e&&e.kind===n.SyntaxKind.ExportDeclaration,isExportAssignment:e=>e&&e.kind===n.SyntaxKind.ExportAssignment,isNamedExports:e=>e&&e.kind===n.SyntaxKind.NamedExports,isVariableStatement:e=>e&&e.kind===n.SyntaxKind.VariableStatement,isModuleDeclaration:e=>e&&e.kind===n.SyntaxKind.ModuleDeclaration},v=(e,t)=>{const r=[];let i=0;return function o(s){!function(t){r.push(e.text.slice(i,t.pos)),i=t.pos}(s);const l=t(s);void 0!==l?(r.push(l),function(e){i=e.end}(s)):(0,n.forEachChild)(s,o)}(e),r.push(e.text.slice(i)),r.filter(Boolean)},g=e=>i(void 0,void 0,void 0,(function*(){var t;const r=yield(0,d.readFile)(e,{encoding:"utf8"}),i=(0,n.parseConfigFileTextToJson)(e,r);if(i.error)throw x([i.error]);const o=i.config,s=(0,n.parseJsonConfigFileContent)(o,n.sys,(0,u.getDir)(e));if(null===(t=s.errors)||void 0===t?void 0:t.length)throw x(s.errors);return{fileNames:s.fileNames,compilerOptions:s.options}})),x=e=>{const t=["Declaration generation failed"];e.forEach((e=>{const r="string"==typeof e.messageText?e.messageText:e.messageText.messageText;if(e.file){const i=e.file.getLineAndCharacterOfPosition(e.start||0);t.push(`${e.file.fileName}(${i.line+1},${i.character+1}): error TS${e.code}: ${r}`)}else t.push(`error TS${e.code}: ${r}`)}));const r=new Error(t.join("\n"));return r.name="EmitterError",r},w=e=>e.replace(/(\.d)?\.ts|\.js$/,"")},584:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.getLocalPackagePath=t.removeExt=t.merge=t.resolve=t.resolver=t.getName=t.getDir=t.normalize=t.PathResolver=void 0;const i=r(17);class o{constructor(e){this.rootPath=(0,t.normalize)((0,i.resolve)(e))}relative(e){return(0,t.normalize)((0,i.relative)(this.rootPath,(0,t.normalize)(e)))}relativeList(e){return e.map((e=>this.relative(e)))}includes(e){return 0===(0,t.normalize)(e).indexOf(this.rootPath)}resolve(...e){return(0,t.normalize)((0,i.resolve)(this.rootPath,...e.filter(Boolean).map((e=>e.replace(/^\/+/,"")))))}resolveList(e){return e.map((e=>this.resolve(e)))}dir(){return(0,t.resolver)((0,t.getDir)(this.rootPath))}res(...e){return(0,t.resolver)(this.resolve(...e))}}t.PathResolver=o;t.normalize=e=>(null==e?void 0:e.replace(/\\/g,"/"))||"";t.getDir=e=>(0,t.normalize)(e).replace(/\/[^/]+\/?$/,"");t.getName=e=>(0,i.basename)((0,t.normalize)(e));t.resolver=e=>new o(e);t.resolve=e=>(0,t.normalize)((0,i.resolve)(e));t.merge=(...e)=>(0,t.normalize)((0,i.join)(...e));t.removeExt=e=>null==e?void 0:e.replace(/\.([^/]+)$/,"");const s=/([^/.]+\/[^/.]+)/;t.getLocalPackagePath=e=>{if(e.includes("node_modules")||e.includes(".yarn"))return"";const t=e.match(s);return t?t[1]:""}},285:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.mkdirp=t.createWriteStream=t.writeJSON=t.readFile=t.readJSON=t.pathExists=t.copy=t.stat=void 0;var i=r(470);Object.defineProperty(t,"stat",{enumerable:!0,get:function(){return i.stat}}),Object.defineProperty(t,"copy",{enumerable:!0,get:function(){return i.copy}}),Object.defineProperty(t,"pathExists",{enumerable:!0,get:function(){return i.pathExists}}),Object.defineProperty(t,"readJSON",{enumerable:!0,get:function(){return i.readJSON}}),Object.defineProperty(t,"readFile",{enumerable:!0,get:function(){return i.readFile}}),Object.defineProperty(t,"writeJSON",{enumerable:!0,get:function(){return i.writeJSON}}),Object.defineProperty(t,"createWriteStream",{enumerable:!0,get:function(){return i.createWriteStream}}),Object.defineProperty(t,"mkdirp",{enumerable:!0,get:function(){return i.mkdirp}})},470:e=>{e.exports=require("fs-extra")},828:e=>{e.exports=require("minimatch")},7:e=>{e.exports=require("typescript")},361:e=>{e.exports=require("events")},17:e=>{e.exports=require("path")}},t={};var r=function r(i){var o=t[i];if(void 0!==o)return o.exports;var s=t[i]={exports:{}};return e[i].call(s.exports,s,s.exports,r),s.exports}(574);module.exports=r})();
