--- packages/builder/src/plugins/DtsPlugin/index.js
@@ -1,1 +1,0 @@
-/*! Copyright (c) 2023 ABux. Under MIT license found in the LICENSE file */(()=>{"use strict";var e={584:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.getLocalPackagePath=t.removeExt=t.merge=t.resolve=t.resolver=t.getName=t.getDir=t.normalize=t.PathResolver=void 0;const o=r(17);class i{constructor(e){this.rootPath=(0,t.normalize)((0,o.resolve)(e))}relative(e){return(0,t.normalize)((0,o.relative)(this.rootPath,(0,t.normalize)(e)))}relativeList(e){return e.map((e=>this.relative(e)))}includes(e){return 0===(0,t.normalize)(e).indexOf(this.rootPath)}resolve(...e){return(0,t.normalize)((0,o.resolve)(this.rootPath,...e.filter(Boolean).map((e=>e.replace(/^\/+/,"")))))}resolveList(e){return e.map((e=>this.resolve(e)))}dir(){return(0,t.resolver)((0,t.getDir)(this.rootPath))}res(...e){return(0,t.resolver)(this.resolve(...e))}}t.PathResolver=i;t.normalize=e=>(null==e?void 0:e.replace(/\\/g,"/"))||"";t.getDir=e=>(0,t.normalize)(e).replace(/\/[^/]+\/?$/,"");t.getName=e=>(0,o.basename)((0,t.normalize)(e));t.resolver=e=>new i(e);t.resolve=e=>(0,t.normalize)((0,o.resolve)(e));t.merge=(...e)=>(0,t.normalize)((0,o.join)(...e));t.removeExt=e=>null==e?void 0:e.replace(/\.([^/]+)$/,"");const n=/([^/.]+\/[^/.]+)/;t.getLocalPackagePath=e=>{if(e.includes("node_modules")||e.includes(".yarn"))return"";const t=e.match(n);return t?t[1]:""}},285:(e,t,r)=>{Object.defineProperty(t,"__esModule",{value:!0}),t.mkdirp=t.createWriteStream=t.writeJSON=t.readFile=t.readJSON=t.pathExists=t.copy=t.stat=void 0;var o=r(470);Object.defineProperty(t,"stat",{enumerable:!0,get:function(){return o.stat}}),Object.defineProperty(t,"copy",{enumerable:!0,get:function(){return o.copy}}),Object.defineProperty(t,"pathExists",{enumerable:!0,get:function(){return o.pathExists}}),Object.defineProperty(t,"readJSON",{enumerable:!0,get:function(){return o.readJSON}}),Object.defineProperty(t,"readFile",{enumerable:!0,get:function(){return o.readFile}}),Object.defineProperty(t,"writeJSON",{enumerable:!0,get:function(){return o.writeJSON}}),Object.defineProperty(t,"createWriteStream",{enumerable:!0,get:function(){return o.createWriteStream}}),Object.defineProperty(t,"mkdirp",{enumerable:!0,get:function(){return o.mkdirp}})},163:function(e,t,r){var o=this&&this.__awaiter||function(e,t,r,o){return new(r||(r=Promise))((function(i,n){function s(e){try{l(o.next(e))}catch(e){n(e)}}function a(e){try{l(o.throw(e))}catch(e){n(e)}}function l(e){var t;e.done?i(e.value):(t=e.value,t instanceof r?t:new r((function(e){e(t)}))).then(s,a)}l((o=o.apply(e,t||[])).next())}))};Object.defineProperty(t,"__esModule",{value:!0}),t.DtsPlugin=void 0;const i=r(903),n=r(285),s=r(780),a=r(584);let l=0;t.DtsPlugin=class{constructor(e){this.path=(0,a.resolver)(e)}apply(e){let t=[];e.hooks.beforeCompile.tapPromise("[dts] start collecting built modules",(()=>o(this,void 0,void 0,(function*(){t=[]})))),e.hooks.compilation.tap("[dts] setup compilation",(e=>{e.hooks.succeedModule.tap("[dts] collect built module",(e=>{if("NormalModule"!==e.constructor.name)return;const r=this.path.relative(e.context||""),o=(0,a.getLocalPackagePath)(r);o&&!t.includes(o)&&t.push(o)}))})),e.hooks.afterCompile.tap("[dts] generate definitions",(()=>{Promise.all(t.map((e=>o(this,void 0,void 0,(function*(){const t=l+=1;try{const r=yield(0,n.readJSON)(this.path.resolve(e,"package.json")),o=r.types,l=r.name,c=this.path.resolve(e),u=this.path.resolve(e,o),p=r.main||"index",d=r.files||"";if(!o)return;let h=this.path.resolve(e,"tsconfig.json");if((yield(0,n.pathExists)(h))||(h=this.path.resolve("tsconfig.json")),!(yield(0,n.pathExists)(h)))return void s.loggers.warn(this.log(t,l,"generation ignored, required tsconfig"));const g=new i.Dts,m=d.map((t=>(0,a.resolver)(c).relative(this.path.resolve(e,t))));g.on("log",(e=>{s.loggers.progress(this.log(t,e))})),s.loggers.info(this.log(t,l,"generation started")),yield g.generate({projectPath:h,name:l,inputDir:c,outputPath:u,main:(0,a.removeExt)(p.replace(/^(\.\/?)+/,"")),filePatterns:m}),s.loggers.success(this.log(t,l,"declaration at",o))}catch(e){s.loggers.error(this.log(t,e.message))}})))))}))}log(e,...t){let r=t.join(" ");return 0!==r.indexOf("[dts")&&(r=`[dts] ${r}`),r.replace(/^\[(dtsw?)\]/,((t,r)=>s.loggers.color(`[${r} ${e}]`,e)))}}},780:e=>{e.exports=require("@abux/logger/cli")},470:e=>{e.exports=require("fs-extra")},903:e=>{e.exports=require("@abux/builder/src/lib/dts/index.js")},17:e=>{e.exports=require("path")}},t={};var r=function r(o){var i=t[o];if(void 0!==i)return i.exports;var n=t[o]={exports:{}};return e[o].call(n.exports,n,n.exports,r),n.exports}(163);module.exports=r})();
