+++ packages/dts/README.md
@@ -1,0 +1,4 @@
+@abux/dts
+=====
+Simple and fast typescript declaration (dts) file generator
+
