+++ packages/dts/lib/DtsWriter.ts
@@ -1,9 +1,6 @@
 /*! Copyright (c) 2023 ABux. Under MIT license found in the LICENSE file */
 import {
   forEachChild,
-  sys,
-  parseConfigFileTextToJson,
-  parseJsonConfigFileContent,
   SyntaxKind,
   createCompilerHost,
   createProgram,
@@ -11,182 +8,17 @@ import {
   ScriptTarget
 } from 'typescript'
 import type {
-  NamedExports,
   SourceFile,
   Node,
-  Diagnostic,
   CompilerOptions,
-  ImportDeclaration,
-  ExternalModuleReference,
-  StringLiteral,
-  ExportAssignment,
-  ExportDeclaration,
-  ModuleDeclaration,
-  LiteralExpression,
-  VariableStatement
+  LiteralExpression
 } from 'typescript'
-
-import { minimatch } from 'minimatch'
+import { type IPathResolver, resolver, normalize, resolve, merge, getDir } from '@abux/paths'
 import EventEmitter from 'events'
-import { type WriteStream, createWriteStream, readFile, pathExists, mkdirp } from '../vendors'
-
-import { getDir, merge, normalize, resolve, resolver } from '../paths'
-import { type IPathResolver } from '../../interfaces'
-
-const EOL = '\n'
-const DTSLEN = '.d.ts'.length
-
-interface IGenerateOptions {
-  // name of package
-  name: string
-  // input directory to list files
-  inputDir: string
-  // optional, main module, default is /index
-  main?: string
-  // path to locate tsconfig.json file
-  // if it's not located in inputDir
-  projectPath?: string
-  // optional, path to place index.d.ts file
-  // if omitted, index.d.ts file will be placed inside input dir
-  outputPath?: string
-  // optional, list of files
-  // if omitted, the full list of files will be loaded
-  // from parsed ts configuration
-  files?: string[]
-  // external types or path references
-  references?: string[]
-  // only emit files matching patterns
-  filePatterns?: string[]
-}
-
-export class Dts extends EventEmitter {
-  async generate ({
-    name,
-    main,
-    inputDir,
-    projectPath,
-    outputPath,
-    files = [],
-    references = [],
-    filePatterns = []
-  }: IGenerateOptions) {
-    let compilerOptions: CompilerOptions = {}
-    const inDir = resolver(inputDir)
-    const outDir = outputPath ? resolver(outputPath).dir() : inDir
-
-    outputPath = outputPath || inDir.resolve('index.d.ts')
-
-    this.emit('log', `[dts] generate for ${inputDir} > ${outputPath}`)
-
-    /* following tsc behaviour, if a project is specified, or if no files are specified then
-    * attempt to load tsconfig.json */
-    if (!files.length || projectPath) {
-      // if project isn't specified, use baseDir.  If it is and it's a directory,
-      // assume we want tsconfig.json in that directory.  If it is a file, though
-      // use that as our tsconfig.json.  This allows for projects that have more
-      // than one tsconfig.json file.
-      const tsConfig = await this.getTsConfig(inputDir, projectPath)
-
-      files = tsConfig.fileNames
-      compilerOptions = tsConfig.compilerOptions
-    }
-
-    // use input values if tsconfig leaves any of these undefined.
-    // this is for backwards compatibility
-    compilerOptions.declaration = true
-    compilerOptions.target = compilerOptions.target || ScriptTarget.Latest // is this necessary?
-    // compilerOptions.moduleResolution = compilerOptions.moduleResolution
-    compilerOptions.outDir = compilerOptions.outDir || outDir.rootPath
-
-    const writeInputDirPath = inDir.rootPath || compilerOptions.rootDir || projectPath
-
-    if (!writeInputDirPath) {
-      throw Error('[dts] input dir path is required')
-    }
-
-    // TODO should compilerOptions.baseDir come into play?
-    const writeInputDir = resolver(writeInputDirPath)
-    const writeOutputDir = outDir.rootPath || compilerOptions.outDir
-    const generatedFiles = files
-
-    const params = [
-      `baseDir = "${writeInputDir.rootPath}"`,
-      `target = ${compilerOptions.target.toString()}`,
-      `outDir = ${writeOutputDir || ''}`,
-      `rootDir = ${compilerOptions.rootDir || ''}`,
-      `moduleResolution = ${compilerOptions.moduleResolution?.toString() || ''}`,
-      'files =',
-      ...generatedFiles.map(file => `  ${file}`)
-    ]
-
-    this.emit(
-      'log:verbose',
-      '[dts] params:\n' + params.map(p => `  ${p}`).join('\n')
-    )
-
-    await mkdirp(getDir(outputPath))
-    const writer = new DtsFilterWriter({
-      outputPath,
-      name,
-      main,
-      references
-    }, {
-      filePatterns
-    })
-
-    writer.on('log', msg => this.emit('log', msg))
-    writer.on('log:verbose', msg => this.emit('log:verbose', msg))
-
-    await writer.write(
-      writeInputDir.rootPath,
-      compilerOptions,
-      generatedFiles
-    )
-  }
-
-  private async getTsConfig (inputDir: string, projectPath?: string) {
-    const inDir = resolver(inputDir)
-    const tsconfigFiles = [
-      projectPath && resolver(projectPath).resolve('tsconfig.json'),
-      projectPath, // if projectPath is location of tsconfig.json
-      inDir.resolve('tsconfig.json')
-    ].filter(Boolean) as string[]
-
-    for await (const tsconfigFile of tsconfigFiles) {
-      if (await pathExists(tsconfigFile)) {
-        this.emit('log', `[dts] tsconfig from ${tsconfigFile}`)
-
-        return await parseTsConfig(tsconfigFile)
-      }
-    }
-
-    throw Error(`Can't find tsconfig in ${projectPath || inDir.rootPath}`)
-  }
-}
-
-interface IDtsWriterOptions {
-  name: string
-  main?: string
-  outputPath: string
-
-  references?: string[]
-  // exclude files
-  excludedPatterns?: string[]
-  // return empty string '' to ignore module
-  resolvedModule?: (resolution: IModuleResolution) => string
-  // return empty string '' to ignore module
-  resolvedImport?: (resolution: IModuleImportResolution) => string
-}
-
-interface IModuleResolution {
-  currentModule: string
-}
-
-interface IModuleImportResolution {
-  currentModule: string
-  importedModule: string
-  isExternal?: boolean
-}
+import { type WriteStream, createWriteStream } from './fs'
+import { minimatch } from 'minimatch'
+import { NodeKinds, getTsError, removeExtension, DTSLEN, EOL, processTree } from './helpers'
+import type { IModuleResolution, IDtsWriterOptions, IModuleImportResolution } from './interfaces'
 
 export class DtsWriter extends EventEmitter {
   readonly ident = '  '
@@ -598,241 +430,3 @@ export class DtsWriter extends EventEmitter {
     this.emit('done')
   }
 }
-
-interface IDtsFilters {
-  filePatterns?: string[]
-}
-export class DtsFilterWriter extends DtsWriter {
-  readonly modulePathMap: Record<string, string> = {}
-  readonly moduleDepsMap: Record<string, string[]> = {}
-  readonly cachedOutputs: Record<string, string> = {}
-
-  constructor (options: IDtsWriterOptions, readonly filters: IDtsFilters) {
-    super(options)
-  }
-
-  protected done () {
-    this.emitOutput()
-    super.done()
-  }
-
-  emitOutput () {
-    const filePatterns = this.filters.filePatterns?.map(p => `**/${removeExtension(p)}`) || []
-    const moduleIds = Object.keys(this.cachedOutputs)
-    const modulePaths = moduleIds.map(id => this.modulePathMap[id])
-    let emittedModuleIds = moduleIds
-
-    if (filePatterns.length) {
-      emittedModuleIds = []
-      modulePaths.forEach((path, idx) => {
-        if (path && filePatterns.some(pattern => minimatch(path, pattern))) {
-          emittedModuleIds.push(moduleIds[idx])
-        }
-      })
-
-      // collect module deps along with module ids
-      emittedModuleIds = emittedModuleIds.map(id => [id].concat(this.collectModuleDeps(id)))
-        .flat()
-        .filter((id, idx, ids) =>
-          ids.indexOf(id) === idx && // de-duplicate
-          moduleIds.includes(id) // only take modules with declarations, since deps added
-        )
-        .reverse()
-    }
-
-    const main = this.resolvedMainId
-
-    if (main && !emittedModuleIds.includes(this.options.name)) {
-      emittedModuleIds.push(this.options.name)
-    }
-
-    emittedModuleIds.forEach(moduleId => {
-      this.emit('log', `[dtsw] declared module ${moduleId}`)
-      this.writeOutput(this.cachedOutputs[moduleId])
-    })
-  }
-
-  private readonly collectModuleDeps = (resolvedModuleId: string): string[] => {
-    const path = this.modulePathMap[resolvedModuleId]
-
-    if (!path) {
-      return []
-    }
-
-    const deps = this.moduleDepsMap[path] || []
-    const depDeps = deps.map(depModuleId => this.collectModuleDeps(depModuleId)).flat()
-
-    return deps.concat(depDeps)
-  }
-
-  protected writeOutputModule (moduleId: string, contents: string) {
-    const lines = this.filterOutput(contents)
-
-    if (!lines.length) {
-      return
-    }
-
-    const output = [
-      `declare module '${moduleId}' {`,
-      this.deduplicateLines(lines).join(EOL),
-      '}'
-    ].join(EOL)
-
-    if (this.filters.filePatterns?.length) {
-      // delay output until the end (done())
-      this.cachedOutputs[moduleId] = output
-
-      return
-    }
-
-    this.writeOutput(output)
-  }
-
-  protected resolveModule (resolution: IModuleResolution): string {
-    const resolvedModuleId = super.resolveModule(resolution)
-
-    this.modulePathMap[resolvedModuleId] = resolution.currentModule
-
-    return resolvedModuleId
-  }
-
-  protected resolveImport (resolution: IModuleImportResolution): string {
-    const resolvedModuleId = super.resolveImport(resolution)
-
-    if (!this.moduleDepsMap[resolution.currentModule]) {
-      this.moduleDepsMap[resolution.currentModule] = []
-    }
-
-    this.moduleDepsMap[resolution.currentModule].push(resolvedModuleId)
-
-    return resolvedModuleId
-  }
-}
-
-const NodeKinds = {
-  isDeclareKeyWord (node: Node): node is ImportDeclaration {
-    return node && node.kind === SyntaxKind.DeclareKeyword
-  },
-  isImportDeclaration (node: Node): node is ImportDeclaration {
-    return node && node.kind === SyntaxKind.ImportDeclaration
-  },
-  isExternalModuleReference (node: Node): node is ExternalModuleReference {
-    return node && node.kind === SyntaxKind.ExternalModuleReference
-  },
-  isStringLiteral (node: Node): node is StringLiteral {
-    return node && node.kind === SyntaxKind.StringLiteral
-  },
-  isExportDeclaration (node: Node): node is ExportDeclaration {
-    return node && node.kind === SyntaxKind.ExportDeclaration
-  },
-  isExportAssignment (node: Node): node is ExportAssignment {
-    return node && node.kind === SyntaxKind.ExportAssignment
-  },
-  isNamedExports (node: Node): node is NamedExports {
-    return node && node.kind === SyntaxKind.NamedExports
-  },
-  isVariableStatement (node: Node): node is VariableStatement {
-    return node && node.kind === SyntaxKind.VariableStatement
-  },
-  isModuleDeclaration (node: Node): node is ModuleDeclaration {
-    return node && node.kind === SyntaxKind.ModuleDeclaration
-  }
-}
-
-type IReplacer = (node: Node) => string | undefined
-const processTree = (sourceFile: SourceFile, replacer: IReplacer): string[] => {
-  const codes: string[] = []
-  let cursorPosition = 0
-
-  function skip (node: Node) {
-    cursorPosition = node.end
-  }
-
-  function readThrough (node: Node) {
-    codes.push(sourceFile.text.slice(cursorPosition, node.pos))
-    cursorPosition = node.pos
-  }
-
-  function visit (node: Node) {
-    readThrough(node)
-
-    const replacement = replacer(node)
-
-    if (replacement !== undefined) {
-      codes.push(replacement)
-      skip(node)
-    } else {
-      forEachChild(node, visit)
-    }
-  }
-
-  visit(sourceFile)
-  codes.push(sourceFile.text.slice(cursorPosition))
-
-  return codes.filter(Boolean)
-}
-
-/**
- * Load and parse a TSConfig File
- * @param options The dts-generator options to load config into
- * @param fileName The path to the file
- */
-const parseTsConfig = async (fileName: string): Promise<{
-  fileNames: string[]
-  compilerOptions: CompilerOptions
-}> => {
-  const configText = await readFile(fileName, { encoding: 'utf8' })
-  const result = parseConfigFileTextToJson(fileName, configText)
-
-  if (result.error) {
-    throw getTsError([result.error])
-  }
-  const configObject = result.config
-  const configParseResult = parseJsonConfigFileContent(configObject, sys, getDir(fileName))
-
-  if (configParseResult.errors?.length) {
-    throw getTsError(configParseResult.errors)
-  }
-
-  return {
-    fileNames: configParseResult.fileNames,
-    compilerOptions: configParseResult.options
-  }
-}
-
-/**
- * A helper that takes TypeScript diagnostic errors and returns an error
- * object.
- * @param diagnostics The array of TypeScript Diagnostic objects
- */
-const getTsError = (diagnostics: Diagnostic[]) => {
-  const messages = ['Declaration generation failed']
-
-  diagnostics.forEach(diagnostic => {
-    const messageText = typeof diagnostic.messageText === 'string'
-      ? diagnostic.messageText
-      : diagnostic.messageText.messageText
-
-    // not all errors have an associated file: in particular, problems with a
-    // the tsconfig.json don't; the messageText is enough to diagnose in those
-    // cases.
-    if (diagnostic.file) {
-      const position = diagnostic.file.getLineAndCharacterOfPosition(diagnostic.start || 0)
-
-      messages.push(
-        `${diagnostic.file.fileName}(${position.line + 1},${position.character + 1}): ` +
-        `error TS${diagnostic.code}: ${messageText}`
-      )
-    } else {
-      messages.push(`error TS${diagnostic.code}: ${messageText}`)
-    }
-  })
-
-  const error = new Error(messages.join('\n'))
-
-  error.name = 'EmitterError'
-
-  return error
-}
-
-const removeExtension = (filePath: string) => filePath.replace(/(\.d)?\.ts|\.js$/, '')
