+++ packages/dts/lib/fs.ts
@@ -1,0 +1,11 @@
+export {
+  stat,
+  copy,
+  pathExists,
+  readJSON,
+  readFile,
+  writeJSON,
+  type WriteStream,
+  createWriteStream,
+  mkdirp
+} from 'fs-extra'
