+++ packages/dts/package.json
@@ -1,0 +1,42 @@
+{
+  "name": "@abux/dts",
+  "description": "Simple and fast typescript declaration (dts) file generator",
+  "version": "0.0.1",
+  "types": "index.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/abuxvn/source.git",
+    "directory": "packages/dts"
+  },
+  "keywords": [
+    "abux",
+    "typescript",
+    "declaration",
+    "dts",
+    "fast",
+    "generation",
+    "generate",
+    "generator"
+  ],
+  "author": {
+    "name": "Hung Luu",
+    "email": "hungluu2106@gmail.com"
+  },
+  "bugs": {
+    "url": "https://github.com/abuxvn/source/issues?q=is%3Aopen+is%3Aissue+label%3A%22package%3Adts%22"
+  },
+  "homepage": "https://github.com/abuxvn/source/tree/main/packages/dts#readme",
+  "files": [
+    "./index.js",
+    "./index.d.ts"
+  ],
+  "dependencies": {
+    "@abux/paths": "workspace:^",
+    "fs-extra": "^11.1.1",
+    "minimatch": "^9.0.2",
+    "typescript": "^5.0.2"
+  },
+  "devDependencies": {
+    "@types/fs-extra": "^11.0.1"
+  }
+}
