+++ packages/paths/README.md
@@ -1,0 +1,3 @@
+@abux/paths
+=====
+A cross-platform path resolving utility used internally by ABux packages
