+++ packages/paths/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "@abux/paths",
+  "version": "0.0.1",
+  "description": "A cross-platform path resolving utility used internally by ABux packages",
+  "types": "index.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/abuxvn/source.git",
+    "directory": "packages/paths"
+  },
+  "keywords": [
+    "abux",
+    "paths",
+    "resolver"
+  ],
+  "author": {
+    "name": "Hung Luu",
+    "email": "hungluu2106@gmail.com"
+  },
+  "bugs": {
+    "url": "https://github.com/abuxvn/source/issues?q=is%3Aopen+is%3Aissue+label%3A%22package%3Apaths%22"
+  },
+  "homepage": "https://github.com/abuxvn/source/tree/main/packages/paths#readme",
+  "files": [
+    "./index.js",
+    "./index.d.ts"
+  ]
+}
