+++ packages/webpack-dts/index.ts
@@ -1,0 +1,1 @@
+export { WebpackDtsPlugin } from './lib/WebpackDtsPlugin'
