+++ packages/webpack-dts/package.json
@@ -1,0 +1,43 @@
+{
+  "name": "@abux/webpack-dts",
+  "description": "Webpack plugin for typescript declaration dts generation",
+  "version": "0.0.1",
+  "types": "index.d.ts",
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/abuxvn/source.git",
+    "directory": "packages/webpack-dts"
+  },
+  "keywords": [
+    "abux",
+    "webpack",
+    "typescript",
+    "declaration",
+    "dts",
+    "generation",
+    "generate",
+    "generator"
+  ],
+  "author": {
+    "name": "Hung Luu",
+    "email": "hungluu2106@gmail.com"
+  },
+  "bugs": {
+    "url": "https://github.com/abuxvn/source/issues?q=is%3Aopen+is%3Aissue+label%3A%22package%3Awebpack-dts%22"
+  },
+  "homepage": "https://github.com/abuxvn/source/tree/main/packages/webpack-dts#readme",
+  "files": [
+    "./index.js",
+    "./index.d.ts"
+  ],
+  "dependencies": {
+    "@abux/dts": "workspace:^",
+    "@abux/logger": "workspace:^",
+    "@abux/paths": "workspace:^",
+    "fs-extra": "^11.1.1"
+  },
+  "devDependencies": {
+    "@types/fs-extra": "^11.0.1",
+    "@types/webpack": "^5.28.1"
+  }
+}
